プロフィール項目のウィジェットをファイルに移動
